Disabled "gridding" (resizing in steps quantised to the font size) within the editor and adjusted the Any-Configure event handler.
This means that the editor can now display half a character, or that
the vertical scrollbar grows/shrinks to fit in an exact number of
characters.

However it does cure some horrible resizing bugs when running on
RedHat9 using metacity (I believe the problem is with metacity, not
redhat9 or gap4) and generally resizing has been buggy for a long
while. Hopefully this produces fewer problems than it solves.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden@338 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1