Changed the HashFile format slightly. It's now format 1.00.
The key difference is that it has a file footer pointing back to the
hashfile header (so the hashfile can be appended to an archive) and it
also has an offset in the header to apply to all seeks within the
archive itself, so it can be prepending to an archive that's already
been indexed without breaking the offsets.

Extended the hash_tar program to allow control over these header options.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden@822 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1