Fixed a bug in mfflush whereby it could attempt to write HUGE amounts of data (-ve size) when files are truncated before flushing; it now fseeks before doing the write and checks if the size is +ve.
Also fixed mfwrite to correctly reset the flush_pos record.

Added a mode field to the mFILE structure so we can keep track of
append and read-only flags. These are checked for in the mfwrite
function so mfwrite now writes to the correct location when append
mode is used (ie forced to the end of file) and it now returns 0 when
attempting to write to a read-only mFILE.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden@992 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1