New file to implement an experimental automatic assembly breaker function.
Specifically it looks for common "words" (in either sequences or
consensus) to identify repetitive sequence. It then uses these to
clip the ends of sequences to identify whether contigs are held
together entirely by sequence overlaps of repetitive elements. When
found it then verifies this "gap" by checking for valid vs invalid
read-pairs.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden@1046 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1