Fixed an error when passing in fully qualified pathnames. We now chdir() to the directory containing the seq.txt file and work from there.
Also some functions involved in supporting fastq files with
callibrated confidence values. This is unfinished and needs more work,
specifically it doesn't do anything with the sequence/qual yet (just
parses it) and the entire operation should probably work from the
GERALD directory instead of the Bustard directory. Hence for now the
-qf and -qr options are undocumented.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden@1193 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1