MAJOR BUG FIX!
Fixed a bug in reorder_ztr() whereby the sorted order of multiple
chunks of the same chunk type were not "stable". The result of this is
that 3 SMP4 chunks (say A, B, C) may end up sorted A, B, C with
nchunks==9 and C, A, B with nchunks==15. Given that an optimisation
means that we change the number of chunks depending on whether we've
encoded HUFF chunks this causes a "corruption" in as far as the
correct data is stored but with potentially an incorrect meta-data
block for the first SMP4 chunk.

See srf_fix.c to reverse this problem.

Also added a warning regarding the -C option and -qf option. These are
inherently incompatible (right now) as purity filtered data is not
calibrated.

Updated version to v1.8


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden@1250 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1