Overhauled the tcl objects somewhat. Contigs and sequence (more to follow) sport an "io" and a "delete" method. This comes about from the realisation that my object system doesn't really fit in with Tcl.
Consider:

if {$condition} {
   set r [$io get_sequence $seq_num]
   puts length=[$r get_length]
}

This this we query the tg_cache to get a sequence object, we create a
Tcl_Obj associated with that too (stored in $r) and register a
*global* function named after the string form of $r. This allows us to
call it with "$r get_length" as technically objects are not classes
and cannot have methods invoked on them in their own right. The names
are globally unique though so there are no clashes.

Hence when $r goes out of scope it destroys the reference to the
object but does not remove the function. We now have to explicitly
call "$r delete" to tidy up otherwise we end up with memory leaks and
global namespace pollution. ie:

if {$condition} {
   set r [$io get_sequence $seq_num]
   puts length=[$r get_length]
   $r delete
}

It's not enough to track reference counts on $r either as the act of
calling it as a function ($r get_length) changes the type from
sequence to cmdName, decrementing the reference. We cannot remove the
memory yet though, so it's not possible to keep track it seems.

	   - -  - -  - -  - -  - -  - -  - -  - -  - -  - -

Perhaps with hindsight we should use a procedural model instead:

if {$condition} {
   set r [g5::io::get_sequence $io $seq_num]
   puts length=[g5::seq::get_length $r]
}

This way we do not create and remove procedures plus our data
automatically goes out of scope and is tidied up for us. The above is
essentially what happens in perl when io is blessed to be in class
g5::io or $r is blessed to be in class g5::seq. I guess incrTcl has an
equivalent mechanism.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden@1278 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1