Initialise new index records to G_INDEX_NEW instead of zero. I don't think this affects anything though.
Removed dead code from g_unlock_views. The heap_free call is already
performed in update_record and so never actually got called in
unlock_views.

Tidied up some of the reading and writing argument checking steps to
permit len==0 now. The logic of this is that it's easier to allocate
and write a zero length block (for say X*Ncontigs when initialising
the contig order array) than to check for the special case of zero length.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden@1365 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1