*MAJOR AND INCOMPATIBLE CHANGE*
First draft of a new data type: SeqBlock.
This implements a block of sequences, rather than a single sequence
object.  The purpose is to improve compression by allowing for
multiple sequences (up to 1024) to be compressed as a single block of
data at the g-layer.

This requires a LOT of smoke and mirrors in the caching layer to get
right, and a few tweaks to allocation of new sequence objects too (in
tg_sequence.c).

Currently it's a mess of code, but the implementation appears to now
be working. The compression achieve is quite amazing; typically
beating bzip2 of the old gap5 database size while still being
faster. The actual compression ratios vary, but 5:1 isn't unexpected.

Note that it works far better with deep assemblies and less well with
very shallow (eg 1000 genomes) data sets. This is partly due to the
inherent compression from having lots of similar (ie overlapping)
strings and partly due to a better ratio of sequences to bins. (Lots
of almost empty bins is inefficient).


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden@1617 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1