Added a flag to distinguish between the first range and subsequent ranges in the contig iterators.
The first range needs to allow objects that overlap the range to be
included while subsequent objects need to only allow objects that
start entirely within the new range (in order to remove duplicates).

This fixes a bug where the first few sequences in a contig were
rejected when exporting to varying file formats.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden@1667 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1