Added a "-r nseqs" option to tg_index to reserve space in the aux file. This is an interim solution for the internal record number (ie not the actual index into the aux) not being a 64-bit number.
Problems currently arise due to packing 1024 sequences into a SeqBlock
record. We simulate individual sequence records by taking the record
number of the SeqBlock and multiplying by 1024, then adding in the
element number. However this means we have 2^21 instead of 2^31
records avaalable in total, meaning 2 million bins is enough to make
it impossible to allocate any sequences unless we ensure that we
reserve the lower record numbers for the sequence/anno blocked
structures.

The correct solution, to be implemented later, would be to use a
64-bit record type.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden@1894 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1