Substantial improvements to undo:
Removed the Redo functionality for now. I may add this back when undo
itself is finished.

Deleting an entire column of data in the consensus now remembers the
old data so undo works (previously it just added pads). To implement
this a new contig obj method was added, pileup, to return a column of
data.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden@1989 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1