Added anno_get_position and anno_get_position2 functions, analogous to the sequence_get_position* funcs. The two share common code so internally they now call the same function, bin_get_item_position(), that takes a type field.
Added a move_annos method of the tcl seq object. This calls the C
sequence_move_annos() function. It's used within the contig editor to
ensure that annotations are moved with a sequence. Still to do: allow
partial moves, in the case of inserting or deleting with a sequence.

Fixed a bug with the bin_for_range function where the database had
smaller bins than the current default minimum bin size. Sometimes we
were putting things in larger bins than we need to.

The Contig Editor Undo command now undoes changes to tag type and tag
comment.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden@2054 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1