g_unlock_views() now sorts the views by record number so that writing out to the aux file is in file offset order.
I'm not sure if this makes a big difference or not, but it seems wise
to not do things in random (or as was actually regularly happen,
reverse) file offset.

Also fixed update_record to check against 0 as well as
G_NO_IMAGE. This prevents (ignored) errors from heap_free(), although
I'm unsure why we get image of zero anyway.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden@2244 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1