MAJOR overhaul of record numbers.
We now must use tg_rec throughout all the code as the data type for
record numbers instead of int or GRec. GRec still exists, but it
should only be within the tgap/g-* code and tgap/tg_iface_g.c
interface. Basically GRec is the data type for the underlying "g"
library, which is still 32-bit, while tg_rec is used throughout gap5
proper for records.

The purpose of this is:
A) Allow for more sequence record numbers. The given bottom 10-bits of
a sequence (or anno) record are an index into a SeqBlock and the upper
bits the SeqBlock record. This previously meant just 21 bits of record
space for SeqBlock or about 2 million records.

B) Allow for future changes of record number to be an ID into other
storage media. Right now they're an index into the g-layer aux file,
but we plan on using them as an index into a bam file if we permit
copy-on-write (or even an index into Arcturus). For bam the obvious
way is to use the record number as a file offset, which therefore
needs more than 2Gb worth of address space.


Change summary:

1) All use of types for records are now tg_rec.
   For Tcl interfaces, we use Tcl_WideInt instead of int and functions
   like Tcl_GetWideInt instead of Tcl_GetInt. (These are defined to be
   at least 64-bit.)
   Use atorec() instead of atoi() when converting from string to
   tg_rec.

2) All uses of %d in printing record IDs are now %"PRIrec" instead.
   (Similarly PRIbtr for BTRec types in b+tree2.c)

3) Removed the option to reserve sequence record numbers from tg_index
   and allocate() function.

4) Changed the format of some data structures. This means the new gap5
   databases are INCOMPATIBLE with the old ones. This gap5 will still
   read old data, but not vice versa. This was required to handle
   64-bit values.

5) Split GDatabase into GDatabase (g-layer) and database_t (gap5
   layer) types. In theory the G data structures are interface
   specific ("g" layer in this case) and shouldn't ever be used by the
   gap5 code. Not important right now, but a necessary split to
   enforce if we ever switch to, say, sqlite as a storage engine.

6) The parse_args functions for tcl interfaces now have a new data
   type ARG_REC, to replace the usual ARG_INT used before.

7) Added index_del interfaces for sequence and contig names. Not
   strictly part of the tg_rec code, but discovered their omission
   during these changes.

8) Bug fix to btree_search for when looking at an item at the current
   root.

9) Removed bin_id field from bin_index_t. It was never filled out
   properly and was essentially the same as rec.

10) Removed rec arrays from SeqBlock and AnnoEleBlock internal
    structures. The arrays were pointless as they were all sequential
    anyway given we have no reallocation system (and they're were not
    written to disc).

11) Bug fix to bin_get_position when recursing over complemented data
    (not part of tg_rec fixes).



git-svn-id: svn://svn.code.sf.net/p/staden/code/staden@2246 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1