Editor searching by sequence name now accepts a prefix match. This required an extra parameter to the b+tree search code.
Also the mechanism of searching is now different as we search both
through a list of names pulled out of the btree (finding the closest
one in the direction we search) and also using a positional iterator
on the contig. This resolves issues where we may have 10,000 sequences
matching a prefix string, but only 5 sequences away is the first
occurence of this so a linear scan would have found the match
faster.

(Potentially we could modify this to allow searching by names even if
no name index exists, albeit at a time cost.)


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden@2247 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1