Adds support for storing ref_start and ref_end coordinates in the bin Range array (BAM/SAM input only). The editor views these coordinates.
It's a proof of concept really to see how inefficient it is to hold an
original reference coordinate for every single reading. (Answer - only
a couple percent overhead.)

I'm shunting it to a branch for now while I experiment with an
alternative approach - single base checkpoints added to the range to
hold a mapping of reference to padded.  Being of length 1 this should
make complementing easier. Plus hopefully it'll be smaller and simpler
to query.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden@2411 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1