Reference base coordinates are now tracked when the input is a SAM/BAM file. This required adjustments to the on-disc file formats for bins, although the code will still read the old format data too.
In the editor window the position box in the bottom left now has a P
or R label next to it. Clicking this toggles padded vs reference
coordinates. It can be adjusted from the settings menu too. When in
Reference mode the contig coordinate line also updates too.

An editor search mode of "reference indel" has been added, to skip
ahead to the next consensus insertion or deletion to the reference. Do
not confuse these with pads - it's specifically where the data has
been removed. Even when all the sequences have a D operator, if they
have all had a * added then the consensus is the correct length and
still has a 1:1 mapping to the reference, meaning there are no missing
bases to find.  A future useful addition though would be to add
markers for consensus pads that align against a reference base
(deletion of zero length maybe?) so we can search for these.

We also need to decide how to fit unpadded coordinates into this
system still.


Implementation wise, a bin range now has an additional type REFPOS. A
refpos object has no actual record, but instead uses these range
fields to hold a mapping to the absolute reference ID and coordinate.

Added nrefpos and nanno to the bin struct in addition to the existing
nseqs field. Much like nseqs these count the number of objects in this
bin and child bins in a recursive manner.

Updated the contig iterators to use nseqs, nrefpos and nanno fields to
short-cut the searches for next/previous item. Basically this is my
"needle in a haystack" algorithm allowing us to rapidly skip forward
to e.g. the next annotation simply by noting which bins have nanno ==
0, and not requiring a linear read of every bin range.



Other passenger updates that appeared in this large update:
- Editor "stripe mode" or "fanfold mode" has reappeared. This is just
  a visual tweak to make visualising columns easier in deep repetitive
  data.

- Fixed contig annotations in SAM to no longer claim to be on
  sequences.

- Relaxed the SAM data out of order check as it's currently triggered
  by alignments starting with a deletion in the cigar string. We need
  a better check for this.

- SAM/BAM reading code now automatically removes duplicates unless
  tg_index -D flag is specified.

- Fixed a break contig bug where tags to the right of the break point
  are on readings that start to the left of the break point were not
  always moved to the left hand contig (when the tags happened to also
  fall into a new bin further to the right).



git-svn-id: svn://svn.code.sf.net/p/staden/code/staden@2435 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1