Identified a bug in bin_add_range and bin_for_range.
When creating new bins as the smallest bin is larger than
io->min_bin_size and we're in a complemented orientation (our parent
or one of its parents has been complemented), then we were setting
offset incorrect. This meant that sequences were put in the incorrect
location. This bug was in bin_for_range().

Also in bin_add_range() we updated the start_used and end_used bin
fields incorrectly for complemented bins.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden@2539 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1