Fixed an error caused in the contig editor whereby edits to data previously edited and then saved could either not be saved correctly or instead accidentally saved when you exit without saving.
More specifically, when using a child I/O editing uses copy-on-write
to take a local writable version of the master I/O record. Saving
flushed this data out and pushed the child I/O record back to the
master I/O, but it didn't reset the lock from r/w to read-only.
Consequentially the next edit to that record did not perform a
copy-on-write again as the child I/O believed it had a writeable
version already.

Corrected this by adding a cache_upgrade(io, ci, G_LOCK_RO) call.
There still a (new?) bug in that the data is not correctly referenced
counted and so edited data wastes cache, however this is simply a
small memory leak rather than a substantial problem with the
possibility to break data.  I'll submit a new patch with corrected
reference counting when I figure out where that problem comes from.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden@2669 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1