Fixed issues with reference counting and making records r/o again.
An earlier commit to fix the issues of contig editors operating on
live data after an initial cache_flush introduced a bug such that
objects shared by multiple editors would erroneously be set back to
read-only copies despite another editor still having a copy-on-write
portion of that data.

This sounds unlikely to happen, but occurs due to seq_blocks being
split between two contigs. The same seq_block can have reads in two
editors, both of which are being actively edited as they form the
halves of a join editor. When we made the join it flushed the changes
to the seq_block from the first editor, making it read-only, and then
died while flushing the changes to the seq_block from the second
editor.

We now keep better track of reference counting when dealing with
multiple child IOs and also changed the cache_rw code to call
cache_dup any time a cached_item is in a parent io rather than simply
relying on a check for ro/rw (which was the original bug I should have
fixed).

In tg_tcl there are also some reference count leaks still. I fixed the
main ones related to sequence editing though with this patch. Consider
the following scenario.

"set s [$io get_sequence 100]" where $io is a child io. This does a
cache_incr(child_io, seq).

"$s replace_base ..." on this object will then call cache_rw() and
replace_base(). The ts->seq pointer gets replaced with a new one as
cache_rw has done a copy-on-write from master to child-io. However the
initial cache_incr from [$io get_sequence] is no longer valid as it's
in another io entirely. I'm not sure what the best general purpose fix
is other than to track which ref count came from where
(messy). Instead I simply cache_decr / cache_rw / cache_incr to move
the $s ref count from master to child i/o.

This fix needs applying to anno and contig objects too, but it's not
so easy as it's unclear whether we'll accidentally lose an object in
the middle.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden@2691 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1