Protect against a crash while joining.
I'm not sure of the circumstances to trigger it, but while handling
the contig registration events for a join, the editor can sometimes
end up trying to run show_cursor on a position where no range data has
been returned. We now gloss over this and refuse to show the cursor
(as we're about to shut down anyway).


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden@2695 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1