Fixed reference count handling in join_contigs().
A user was rarely seeing the right hand contig not being destroyed
after a contig join. The only possible code route I can see for this
happening is cr being freed during one of the plots redrawing via
contig_notify callbacks, and cr memory being reused to point to, say,
a bin object. Then contig_destroy(io, cr->rec) later on fails as the
object isn't valid.

At least I hope that's the cause of that bug... Whatever the case this
is a bug and it's fixed.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden@2811 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1