Fixed sequence orientation bugs.
One is in the editor view where "<<" and ">>" were reported in the
information line incorrectly on complemented contigs.

The other is more serious and in the export to SAM code. It was
incorrectly setting the orientation flags. This mean exporting a
"--> <--" pair in a complemented contig and re-importing turned them
into a "<-- -->" pair, showing up as red in the template display.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden@2934 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1