Added code to range_populate() when sorting by clipped sequence end. If a sequence overlaps the start..end range but after clipping it does not, throw it away so we deal with it later.
The reason for this is it was providing out of order sorting between
one bucket of data in from range_populate and the next bucket of data
in range_populate, causing iterators to give incorrect ordering. This
in turn shows up as contig holes leading to erroneous contig
breaking around sequences with long cutoff regions.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden@2964 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1