Two fixes.
1) In g_read_alloc check that vi.used is not G_NO_IMAGE before
attempting to read. Our previous assumption that vi.used length is
non-zero is not solely enough as the record may have been re-allocated
from an earlier free record in which case only the image has been
reset. (This is perhaps a bug.)

2) When writing contig links we now forcibly check that data is within
the desired range, in particular that end1, end2 and orientation are
booleans. Failure to be so means we could have underallocated memory
and leads to a corruption.

This is to work around an unknown bug elsehwere which corrupted this
link data. We need to find and fix that too! (On going work.)



git-svn-id: svn://svn.code.sf.net/p/staden/code/staden@3029 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1