Added read pair filtering option to Find Internal Joins.  Turning this on make FIJ check each match to see how many read pairs span between the two contigs in the neighborhood of the aligned region.  The alignments are only kept if the number of read pairs exceeds a given threshold.
The read pair screen actually runs in two stages.  In the first stage, fij
calls fij_prefilter_repeats on the combined contig list.  This calls
spanning_pairs (the guts of Find Read Pairs) to get a list of contigs
that are linked by the desired number of pairs.  It uses this list to make
a HashTable of contig_pair structs and also to prune any contigs with no
read pair links from the list being searched.  The idea is that
spanning_pairs is fairly quick, so using it to reduce the search space should
save time later.

do_it_fij, compare_b_bulk and align_blocks_bulk all now take the HashTable
of contig pairs.  Any word matches that do not match an item in the HashTable
are filtered out so they do not have to be aligned later.

The second stage of filtering takes place in check_overlap_pairs which
is called by add_fij_overlap and add_fij_overlap_r.  This scans for read
pairs over the aligned region, plus a little extra to account for templates
that extend beyond the alignment. It ensures that any spanning pairs
found are a reasonable distance apart and in the correct orientation for the
library.

Other minor changes are:

 * In buffij, pos1 and pos2 are no longer swapped

 * The code to convert a list of libraries to a HashTable has moved from
   spanning_pairs to create_lib_hash.  This allows the library HashTable
   to be used later on in check_overlap_pairs.

 * fij and do_it_fij now take a fij_arg * parameter to vastly reduce the
   number of separate parameters passed in.

 * Read pair screening options have been added to the FIJ dialog box.



git-svn-id: svn://svn.code.sf.net/p/staden/code/staden@3064 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1