Disable rounding up of bin sizes in extend_root_bin
Rounding up the new bin size to a power of 2 gives a more balanced tree
structure.  Unfortunately it also leads to the bins growing in size very
rapidly under certain conditions during long sequences of joins.  In view
of this, the rounding up code has been disabled for now until we come up
with a better solution.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden@3118 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1