Fixed cache look-up of edview structs.
edview_hash now allows duplicate keys so that it can handle the same contig
being open in multiple views.  edview_destroy has been updated so that it
searches for the correct edview struct when removing items.

Added edview_renumber, which changes the contig record number associated
with an editor (e.g. when joining).  It also updates edview_hash so
that the edview struct is refiled under the new record number.  This fixes
a crash caused by an old edview	being used after a join as it wasn't cleaned
up correctly.

Simplified edview_find.  It can only return one edview pointer, so it returns
the first one it finds for the contig record number.  This is likely to be
the last one that was opened.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden@3266 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1