Observed a database corruption where the same record had been freed twice in a row, forming a tight loop in the free list. This then subsequently causes the same record to be allocated twice in a row, causing mayhem.
This workaround is in two parts.

1) g_free_rec_ (which returns a free record, not frees it) now checks
for the tight loop and corrects the free list if it finds this case,
returning G_NO_REC too (ie create a new record as none are free).

2) g_unlock_views with a view of flag G_VIEW_DELETED now does an
assertion for attempting to delete it twice in a row. I've no idea
where this happens, but if it happens again then the assert should
help to pinpoint where.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden@3363 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1