Improved on the g_check_header() and related functions.
This attempts to spot when the global .g5x aux timestamp changes under
the feet of the current process, indicating that someone was naughty
and either removed the BUSY file to open up a second r/w gap5 or
copied another database version over the top of our own one.

Unfortunately due to filesystem caching it was too slow to spot a lot
of errors. We now use a new filedescriptor each time to do the disk vs
memory comparison and we also explicitly call fsync after updating
this to ensure the time is flushed out.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden@3428 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1