Fix bug where g_check_header used the wrong .aux filename.
g_check_header constructed the wrong name for the .g5x file when the
database name was given with a suffix on the command-line.  This caused it
to incorrectly complain about the database being renamed, and (due to it not
stopping then and checking for the wrong value from gfile->low_level_vector)
also that the timestamp had changed.

g_open_file now stores the actual aux file name in the GFile struct, so
g_check_header can use it later.  g_check_header now calls panic_shutdown()
immediately on failure, and also checks for the right return value from
gfile->low_level_vector.

Took the opportunity to remove various fixed-size buffers when dealing with
filenames, especially in find_db_files and actf_lock.

Fixed an invalid read in open_log_file when log files are enabled and the
database name is very short.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden@3448 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1