Fix bugs in spanning_pairs when run in end_all mode.
When spanning_pairs had to use sequence_get_position to look up pair positions,
it was not filling in all the needed parts of r2, leading to the use of
uninitialized values later on.  It now fills in everything that is needed,
which makes the results much more deterministic.  This only affected end_all
mode, as none of the others needed to look up pairs by this method.

Using sequence_get_position could also bring back links to contigs that were
not in the original list passed to spanning_pairs.  These are now filtered
out.  This also fixes a failed assertion in Find Internal Joins that was
triggered by spanning_pairs returning such unexpected links.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden@3510 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1