Reduced the volume of read-pair entries dumped to the /tmp file.
Previously when we saw Read.1st we recorded it in a hash table and then
when we found Read.2nd we linked it back to Read.1st and recorded in
the temporary file a request to link Read.1st to Read.2nd in a second
pass. (As doing this immediately causes a lot of disk thrashing due to
chimeras, mismatches, genome shuffling etc.)

Now we also make the forward link of Read.1st to Read.2nd if it is
within a bin that is still held in the in-memory cache and is still
locked for read-write (meaning it is yet to be saved/flushed).

On a 3Gb bam file test this reduced the number of entries stored in
the tmp file for 2nd pass from 15.9 million down to 0.6 million. The
effect on speed was slight - only 3% quicker overall.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden@3564 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1