Added back fsync of main .g5d file when flushing.
This will slow things down slightly, depending on how often flushes
occur, but resolves some database corruptions we've been seeing.

Specifically if a user has gap5 open on one machine, makes changes and
saves the DB, then creates a backup of the database on the command
line on a different machine[1] then there is a chance that NFS hasn't
correctrly synced the changes yet. This call forces that to happen.

[1] This happens a lot here due to using load-sharing to launch gap5
on one of several nodes in a cluster.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden@3594 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1