Remove duplicate move_seq code and ensure annotations are in same bin as seq.
The code for MOVE_SEQ in contig_cmd could move sequences to new bins without
ensuring that the annotations were moved as well.  The version in
sequence_move did fix the annotations, so made contig_cmd call this instead
of using its own code.

sequence_move now takes a contig_t ** so it can update the location of
the contig record in the caller should it get checked out.  It also gets a
prototype in tg_sequence.h



git-svn-id: svn://svn.code.sf.net/p/staden/code/staden@3612 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1