Further work on read-group machine types.
1) Bug fix to the tg_index process so that it looks up "PL"atform RG
tag correctly.  Previously it was doing something bizarre with
attempting to find the library name as a 2 character key.

2) Changed seq_tech to be an int instead of int:3 in the seq_t struct
as it has 9 possible values now.

3) Changed the on-disk format, which also packs 3 bits, to steal 2
bits from the parent_type field and use these for seq_tech.  This is
backwards compatible as these top bits have always been zero and we've
also never needed more than 7 for seq_tech either.

This limits the available number of GT_* types to 64, but we're only
using 18 at the moment and I don't see us adding that many more table
types to gap5. It gives us up to 32 instrument types.

4) Added set_name and set_machine_type methods to the library tcl
object.

5) The List Libraries window now allows editing of the library name
too, also also saves both this and edits to library machine type.


Note that editing the machine type here does not actually modify all
the sequence structures as the library type has been cached in both
seq and range structures during tg_index. I do not know how to prevent
that in an efficient manner, so for now editing the type is purely for
informational purposes.



git-svn-id: svn://svn.code.sf.net/p/staden/code/staden@3709 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1