Fixed the mapping quality filter range from 0-80 to 0-255 as we are seeing data with >80 as mquals.
Fixed the use of TEMPLATE_DISPLAY.SETTINGS_$i library records. These
could be for another database, so it already attempted to blank out
the libraries-list when they don't match up, but it failed to clear
the undo versions causing the filter window to possibly blank the data
when it's invoked.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden@3711 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1