Updated the consensus algorithm to cope with different overcall and undercall probabilities per machine platform, permitting mixing of different technologies to form a single consensus.
At present these are complete guesses with only a minimal nod to
reality and have no measurements backing them up (yet).  Expect them
to be tuned in the future. However they act as a good proof of
concept.

The Tcl variable default_seq_tech may be modified to change what the
Unknown sequencing tech is treated as.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden@3715 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1