Allowed the default_seq_tech parameter to be modified from Gap5 GUI.
To do this required several things.

1) The cached consensus may be incorrect after changing this, so it
has to be purged. 

2) The parameter needs to be remembered in the database itself so it
can be set the next time we open a database.  I tidied up the setting
of default parameters in multiple locations and put them into
gap_globals.tcl.

3) To store an arbitrary variable without modifying the database
format required some lateral thinking. WARNING: ugliness ahead. Those
of a nervous disposition should look away now.

We create an anno_ele object with parent_type of GT_Database; ie an
annotation on the database rather than a contig or sequence. 
A scaffold of name __g5::CONFIG is created whose record number points
to this annotation instead of a scaffold.  The annotation contains
a series of 'set key val' tuples, which are carefully checked for
validty so we don't blindly trust everything.

We can extend this with other __g5::* fields for different types of
data, replacing the old Gap4 "Notes" mechanism as tags attached to a
database are essentially the same thing as notes.

4) Added various C/Tcl interfaces to set and get the config_anno
field.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden@3716 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1