First draft of a new haplotype finder.  It replaces the Sort By Sequence mode in the contig editor as well as generating lists of read IDs per haplotype and creating a master "haplotypes" list of lists.
Disassemble readings now knows how to handle list of lists, so that it
disassembles each sub-list into its own set of contigs.

There is more to do on this to tidy it up: handling read-pairs,
optimising it for larger contigs, improving the consensus algorithm to
label more things as SNPs by avoiding the dominance of depth.  However
it's a good starting point and users are already wanting to use it so
we'll see how it goes in the current state.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden@3750 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1