Made contig_insert_base2() more robust with respect to corrupted data or other sources of errors.
Previously it just bailed out of the recursion, leaving a half
inserted column. This is complicated to deal with and often turns a
small problem into a bigger problem.

Now it boldly carries on regardless, completing as much as it can,
before then returning -1 to the calling function to indicate the
failure.  Ideally the calling function would execute this in a child
io so it can be backed out of to revert the change, but in the absence
of this we have at least maintained consistency where able.


git-svn-id: svn://svn.code.sf.net/p/staden/code/staden@3751 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1